Changes in response to code review
