+++ shop/src/themes/shared/checkout/Information.js
@@ -12,6 +12,8 @@ import Link from 'components/Link'
 import CountrySelect from 'components/CountrySelect'
 import ProvinceSelect from 'components/ProvinceSelect'
 
+import determineTaxes from './_Taxes'
+
 export const Information = () => {
   const history = useHistory()
   const [{ cart }, dispatch] = useStateValue()
@@ -32,38 +34,10 @@ export const Information = () => {
           return
         }
         dispatch({ type: 'updateUserInfo', info: newState })
-
-        //Calculate taxes
-        try {
-          if (config.taxRates.length) {
-            const taxedCountry = config.taxRates.find(
-              (obj) => obj.country == Countries[newState.country].code
-            )
-
-            //If taxes are configured on the 'country' level, do this
-            let taxRate = get(taxedCountry, 'rate', 0)
-
-            //If taxes are configured at the 'province' level, do this
-            if (newState.province && taxedCountry.provinces) {
-              const currentProvinceCode = get(
-                Countries[newState.country].provinces,
-                newState.province
-              )['code']
-
-              for (const provinceObject of taxedCountry.provinces) {
-                if (provinceObject.province == currentProvinceCode) {
-                  taxRate = provinceObject.rate
-                  break
-                }
-              }
-            }
-
-            dispatch({ type: 'updateTaxRate', taxRate })
-          }
-        } catch (err) {
-          console.error('Error: Taxes not added to cart', err)
-        }
-
+        dispatch({
+          type: 'updateTaxRate',
+          taxRate: determineTaxes(config, newState)
+        })
         history.push({
           pathname: '/checkout/shipping',
           state: { scrollToTop: true }
@@ -175,6 +149,7 @@ export const MobileInformation = () => {
       return
     }
     dispatch({ type: 'updateUserInfo', info: newState })
+
     history.push('/checkout/shipping-address')
   }
   return (
